terraform-modules/colab-deployment/cloud-function/requirements.txt
functions-framework==3.*
google-cloud-dataform==0.5.5
